security(cef): Windows sandbox Phase 3 — DLL wrapper pattern (issue #1374)
Restructures agentmux-cef as a [lib] cdylib + thin [[bin]] wrapper so
bootstrap.exe (renamed agentmux-cef.exe) can initialize cef_sandbox_info_t
before loading us as a DLL and calling RunWinMain. Removes the
target_os = "windows" exclusion from the no_sandbox guard so Windows +
sandbox feature now activates the renderer sandbox.
