+++ .changesets/1781996547534-security-cef-windows-sandbox-phase3-dll-wrapper-1374-w9kx.md
@@ -1,0 +1,4 @@
+---
+type: patch
+---
+security(cef): enable renderer sandbox on Windows via DLL wrapper pattern (issue #1374)
